requirements.txt
transformers>=4.40.1
torch>=2.2.0
pandas
datasets
scikit-learn
matplotlib
accelerate>=0.21.0
uvicorn[standard]
python-dotenv
fastapi
sqlalchemy
psycopg2-binary
jinja2
python-multipart
# Optional dependencies for specific functionalities